% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{as_summary_text}
\alias{as_summary_text}
\title{Convert rNotify Payload to Formatted Summary Text}
\usage{
as_summary_text(x, ...)
}
\arguments{
\item{x}{An \code{rnotify_payload} object.}

\item{...}{Additional arguments.}
}
\value{
Character string.
}
\description{
Generates clean, human-readable plain text or markdown for push notifications and emails.
}
