% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throttle.R
\name{create_throttler}
\alias{create_throttler}
\title{Create a Throttle Controller}
\usage{
create_throttler(
  throttle_sec = NULL,
  milestones = NULL,
  pct_step = NULL,
  min_step_change = NULL
)
}
\arguments{
\item{throttle_sec}{Minimum interval in seconds between successive notifications.
Default is NULL (no time throttle).}

\item{milestones}{Numeric vector of progress percentage milestones to trigger on,
e.g. \code{c(25, 50, 75, 100)}. Default is NULL.}

\item{pct_step}{Percentage interval to trigger on, e.g. \code{10} for every 10\%. Default is NULL.}

\item{min_step_change}{Minimum integer steps advanced before allowing another progress notification.}
}
\value{
An environment acting as a stateful throttle controller.
}
\description{
Create a Throttle Controller
}
