% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{current_iso8601}
\alias{current_iso8601}
\title{Get Current Timestamp in ISO-8601 Format}
\usage{
current_iso8601()
}
\value{
Formatted ISO-8601 string with local timezone.
}
\description{
Get Current Timestamp in ISO-8601 Format
}
