% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{format.rnotify_payload}
\alias{format.rnotify_payload}
\title{Format Method for rNotify Payload}
\usage{
\method{format}{rnotify_payload}(x, ...)
}
\arguments{
\item{x}{An \code{rnotify_payload} object.}

\item{...}{Additional arguments.}
}
\value{
Formatted character string.
}
\description{
Format Method for rNotify Payload
}
