% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_duration}
\alias{format_duration}
\title{Format Duration in Seconds to HH:MM:SS}
\usage{
format_duration(seconds)
}
\arguments{
\item{seconds}{Numeric elapsed or ETA time in seconds.}
}
\value{
Formatted character string, e.g. "01:23:45".
}
\description{
Format Duration in Seconds to HH:MM:SS
}
