% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_pid_alive}
\alias{is_pid_alive}
\title{Check if an OS Process ID (PID) is Active}
\usage{
is_pid_alive(pid)
}
\arguments{
\item{pid}{Integer process identifier.}
}
\value{
Logical TRUE if process is currently running, FALSE otherwise.
}
\description{
Cross-platform process check.
}
