% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{new_notif}
\alias{new_notif}
\title{New notification}
\usage{
new_notif(title, message, subtitle = NULL, icon = NULL, sound = NULL,
  time = NULL)
}
\arguments{
\item{title}{Title of the notification}

\item{message}{Message of the notification}

\item{subtitle}{Subtitle of the notification (not available on windows).}

\item{icon}{Icon of the notification}

\item{sound}{Sound of the notification}

\item{time}{Time in milliseconds of the notification.}
}
\description{
New notification
}
