% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_notify.R
\name{notify_current_job}
\alias{notify_current_job}
\title{Access the Currently Active Enclosing Job}
\usage{
notify_current_job()
}
\value{
The active \code{NotifyJob} object, or NULL if not within \code{with_notify}.
}
\description{
Access the Currently Active Enclosing Job
}
