% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_notify.R
\name{notify_progress}
\alias{notify_progress}
\title{Report Progress from Inside with_notify Block}
\usage{
notify_progress(step, total = NULL, message = "")
}
\arguments{
\item{step}{Current step number.}

\item{total}{Total steps.}

\item{message}{Status message.}
}
\description{
Report Progress from Inside with_notify Block
}
