% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_base.R
\name{print.rnotify_channel}
\alias{print.rnotify_channel}
\title{Print Method for rNotify Channel}
\usage{
\method{print}{rnotify_channel}(x, ...)
}
\arguments{
\item{x}{An \code{rnotify_channel} object.}

\item{...}{Additional arguments.}
}
\description{
Print Method for rNotify Channel
}
