% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{rnotify_get_configured}
\alias{rnotify_get_configured}
\title{Retrieve Configured Notification Channels}
\usage{
rnotify_get_configured()
}
\value{
Named list of \code{rnotify_channel} objects.
}
\description{
Retrieve Configured Notification Channels
}
