% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rnotify_log}
\alias{rnotify_log}
\title{Internal Logging for rNotify}
\usage{
rnotify_log(msg, level = "INFO")
}
\arguments{
\item{msg}{Character log message.}

\item{level}{Character log level: "INFO", "WARN", "ERROR".}
}
\description{
Appends diagnostic messages to a local log file and optionally displays a warning.
}
\keyword{internal}
