% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel.R
\name{start_sentinel}
\alias{start_sentinel}
\title{Start the Background Sentinel Watcher Process}
\usage{
start_sentinel(
  job_id,
  title,
  channels,
  parent_pid = Sys.getpid(),
  poll_interval = 10
)
}
\arguments{
\item{job_id}{Unique job ID.}

\item{title}{Job title.}

\item{channels}{List of \code{rnotify_channel} objects.}

\item{parent_pid}{Integer PID of the main R session (default: \code{Sys.getpid()}).}

\item{poll_interval}{Integer check interval in seconds (default: 10).}
}
\value{
An environment containing sentinel handles and methods.
}
\description{
Start the Background Sentinel Watcher Process
}
