% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_notify.R
\name{notify_progress}
\alias{notify_progress}
\title{Report Progress from Inside with_notify Block or Directly on Job}
\usage{
notify_progress(step, total = NULL, message = "", job = NULL)
}
\arguments{
\item{step}{Current step number or \code{NotifyJob} object.}

\item{total}{Total steps.}

\item{message}{Status message.}

\item{job}{Optional target \code{NotifyJob} instance.}
}
\description{
Report Progress from Inside with_notify Block or Directly on Job
}
