% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_notify.R
\name{notify_step}
\alias{notify_step}
\title{Report a Step Milestone from Inside with_notify Block}
\usage{
notify_step(message = "")
}
\arguments{
\item{message}{Milestone description.}
}
\description{
Report a Step Milestone from Inside with_notify Block
}
